+++ application/src/main/java/org/opentripplanner/service/streetdecorator/model/VertexLevelInfo.java
@@ -3,7 +3,7 @@ package org.opentripplanner.service.streetdecorator.model;
 import javax.annotation.Nullable;
 
 public record VertexLevelInfo(
-  @Nullable Float floorNumber,
+  @Nullable Double floorNumber,
   @Nullable String name,
   long osmVertexId
 ) {}
